added flag to bandit spawns, first pass on assignTroops rewrite
Added an IsBandit flag to all the spawns in spawnlist_bandits.nut to
make it easy to check for them for testing purposes.

First pass on rewriting assignTroops() in z_world_entity_common.nut.
Kinda clunky but hopefully functional; it uses weighting but doesn't
account for what's already been added, which could lead to weird spawn
groups. Next goal is to do the same to createDefenders in location.nut
and addUnitToEntity in contract.nut.