fixed legend_poison_weapon.nut, modified legend_nightvision.nut
Legend_poison_weapon.nut would crash the game is it was used while the
weapon still had poison uses left; it was calling the wrong event.
Fixed it to the proper one in the poison weapon effect skill.

Changed legend_nightvision.nut to also give night vision to the user;
it seemed strange to me that it doesn't. Feel free to disregard if it
seems unbalanced.