fixed item description, added favored enemies
Legend_named_warlock_hood.nut had mismatching description and tooltips
regarding the extra hitpoints, fixed it to match.
Added Necrosavant Lords and Winselmutters to the favored enemy
Necrosavant perks, to match Necrosavants and Ghosts being a part of
it. Guessed on appropriate values, feel free to change. Still needs to
be tested