Fix Schrat and Conqueror FE
Fix Schrat sapling being missing from Favoured Enemy mapping for bonus per kill. It is worth 0.2% per kill. Also fixed the Greenwood Schrat sapling to grant 0.4% bonus per kill as it says in the description.

Gave The Conqueror (SkeletonBoss) a 5% bonus per kill, since there is only one of it and it's the boss of the Black Monolith.