Adds a simple system to determine the cargo budget for caravans. Depending on the size and type of settlement, the budget can be within this 500-6000 crowns.
Settlements now will invest 2-3% of their wealth to each caravan, and will receive a profit of 1-2% wealth if the caravan arrives at its destination.

Caravan's guards are scaling with the difficult and days now. This should make caravan more risky to raid, still lose to any big army but should be able to put up a good fight against the regular raider.