Adds 'getImportedGoodsInventory' function to "settlement.nut" script, just wondering why i didn't add it before.
Makes sure so that the 'this.m.ImportedGoodsInventory' stash only contains at most 50 items. Adding new items will remove the oldest items. This is because shops in a settlement only update their stash when player enters that settlement, i don't want that settlement to accumulate hundreds of item from trade and then flood the marketplace with an overwhelming amount of item.