extendVarsWithPronouns identiy; Demo: contract
The extendVarsWithPronouns function now takes in an additional optional _characterIdentitySuffix parameter, which helps apply gendered pronouns on a per identity basis.

For example, calling `extendVarsWithPronouns(vars, poacher.getGender(), "poacher")` will mean that any supported pronoun placeholder with the "poacher" suffix will be applied for based on the poacher's gender.

Suppose that poacher.getGender() is male, then:
- "%They_poacher%" -> "He"
- "%person_poacher%" -> "man"
- etc

This is now supported by default for "employer", "randombrother" and "randombrother2" for all contract event text.

For an example of how it is used, see the changes applied to `intro_templates`, `negotiation_templates`, and `roaming_beasts_contract`

These changes provide pronoun support for the contract's employer.