Meek rework - evade first attack each combat
Meek no longer provides a chance to force attackers to reroll. The maths showed that this resulted in very limited benefit. Try drawing a probability tree and observing the difference in probability of landing a successful hit with and without rerolls.

Instead, Meek now grants a guaranteed evade for the first attack made against the character every combat encounter

- New `IsEvadingAllAttacks` value in character `_properties`. When true for the defender, they will ignore all hit calculations and evade the incoming attack.
- Meek uses the new `IsEvadingAllAttacks` property