Item Tooltip shows Trade History
New system to track the history of items - which settlement they were produced at, and which settlements have they been exported through via caravans or the player. (requires World Economy to be enabled)

The information is displayed in the Item tooltip.

New tooltip info can be configured in MSU Mod Setting (default is enabled)