Fix wrong border key in intensive trait tooltip
The key should be `divider` instead of `border`. Doesn't seem to work during local testing, but that is supposed to be the correct syntax.

Either there's some issue with the patch I made or I will need to check the tooltip module to see why the `divider` property is not getting picked up